Fixed 0 Life

+++ Assets/william/script/SO/HpData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New HpData", menuName = "Item/Hp Data", order = 1)]
 public class HpData : GameData
 {
-    public float hp;
-    public float hp_max;
+    public int hp;
+    public int hp_max;
     public bool IsOver()
     {
         return hp <= 0;
